pyproject.toml
[project]
name = "ambilight-iskarna"
version = "0.1.0"
description = "Ambilight-style lighting controller for IKEA ISKARNA via Home Assistant."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "httpx>=0.26.0",
    "dxcam>=0.0.5",
    "Pillow>=10.2.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
    "black>=24.0.0",
    "isort>=5.12.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["ambilight"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
